feat: refactor delete table record
